lambda_parse_pptx_from_s3/requirements.txt
python-pptx==0.6.22
lxml
Pillow